+++ package.json
@@ -15,7 +15,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-dual-range-slider": "^1.0.4",
-    "react-router-dom": "^6.18.0",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "^5.0.1"
   },
   "devDependencies": {
