+++ src/components/DoubleRangeSlider.css
@@ -4,7 +4,7 @@
   align-items: center; /* Center children horizontally in the cross axis */
   justify-content: center; /* Center children vertically in the main axis */
   height: 20vh;
-  width: 100vw;
+  width: auto;
 }
 
 .slider {
