+++ src/components/Slider.jsx
@@ -1,10 +1,13 @@
 import React from 'react';
 import './Slider.css';
 
-export const Slider = ({ label, min, max, value, step, onChange }) => {
+export const Slider = ({ label, min, max, value, step, onChange, formatLabel }) => {
+  // Use formatLabel function if provided, otherwise default to current behavior
+  const displayLabel = formatLabel ? formatLabel(value) : `${value} cm`;
+
   return (
     <div className="slider-container">
-      <label className="slider-label">{label}: {value >= max ? `> ${max} cm` : `${value} cm`}</label>
+      <label className="slider-label">{label}: {displayLabel}</label>
       <input
         type="range"
         min={min}
